--- src/app/chat/[id]/page.tsx
@@ -1,8 +1,0 @@
-import MainChat from './_component/Chat';
-export default async function Chat({ params }: { params: { id: string } }) {
-	return (
-		<>
-			<MainChat />
-		</>
-	);
-}
